+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentViewHolder.java
@@ -1,11 +1,9 @@
 package com.getstream.sdk.chat.adapter;
 
 import android.content.Context;
-import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.text.TextUtils;
 import android.util.Log;
-import android.util.TypedValue;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
@@ -220,13 +218,14 @@ public class AttachmentViewHolder extends BaseAttachmentViewHolder {
     }
 
     private void applyStyle() {
-        boolean isMine = getMessageListItem().isMine();
-        tv_media_title.setTextSize(TypedValue.COMPLEX_UNIT_PX, style.getAttachmentTitleTextSize(isMine));
-        tv_media_title.setTextColor(style.getAttachmentTitleTextColor(isMine));
-        tv_media_title.setTypeface(Typeface.DEFAULT_BOLD, style.getAttachmentTitleTextStyle(isMine));
 
-        tv_media_des.setTextSize(TypedValue.COMPLEX_UNIT_PX, style.getAttachmentDescriptionTextSize(isMine));
-        tv_media_des.setTextColor(style.getAttachmentDescriptionTextColor(isMine));
-        tv_media_des.setTypeface(Typeface.DEFAULT_BOLD, style.getAttachmentDescriptionTextStyle(isMine));
+        if(getMessageListItem().isMine()) {
+            style.attachmentTitleTextMine.apply(tv_media_title);
+            style.attachmentDescriptionTextMine.apply(tv_media_des);
+        } else {
+            style.attachmentTitleTextTheirs.apply(tv_media_title);
+            style.attachmentDescriptionTextTheirs.apply(tv_media_des);
+        }
+
     }
 }