+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentViewHolderFile.java
@@ -2,7 +2,6 @@ package com.getstream.sdk.chat.adapter;
 
 
 import android.content.Context;
-import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.util.TypedValue;
 import android.view.ViewGroup;
@@ -13,6 +12,7 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 
 import com.getstream.sdk.chat.R;
 import com.getstream.sdk.chat.model.Attachment;
+import com.getstream.sdk.chat.utils.TextViewUtils;
 import com.getstream.sdk.chat.view.MessageListView;
 import com.getstream.sdk.chat.view.MessageListViewStyle;
 
@@ -47,14 +47,15 @@ public class AttachmentViewHolderFile extends BaseAttachmentViewHolder {
     }
 
     private void applyStyle() {
-        boolean isMine = getMessageListItem().isMine();
-        tv_file_title.setTextSize(TypedValue.COMPLEX_UNIT_PX, style.getAttachmentTitleTextSize(isMine));
-        tv_file_title.setTextColor(style.getAttachmentTitleTextColor(isMine));
-        tv_file_title.setTypeface(Typeface.DEFAULT_BOLD, style.getAttachmentTitleTextStyle(isMine));
 
-        tv_file_size.setTextSize(TypedValue.COMPLEX_UNIT_PX, style.getAttachmentFileSizeTextSize(isMine));
-        tv_file_size.setTextColor(style.getAttachmentFileSizeTextColor(isMine));
-        tv_file_size.setTypeface(Typeface.DEFAULT_BOLD, style.getAttachmentFileSizeTextStyle(isMine));
+        if(getMessageListItem().isMine()) {
+            style.attachmentTitleTextMine.apply(tv_file_title);
+        } else {
+            style.attachmentTitleTextTheirs.apply(tv_file_title);
+        }
+
+
+        style.attachmentFileSizeText.apply(tv_file_size);
     }
 
     private void configAttachment() {