+++ library/src/main/java/com/getstream/sdk/chat/view/MessageListViewStyle.java
@@ -4,21 +4,18 @@ import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.graphics.drawable.Drawable;
 import android.util.AttributeSet;
 
 import com.getstream.sdk.chat.R;
+import com.getstream.sdk.chat.style.TextStyle;
 
 public class MessageListViewStyle extends BaseStyle {
     private static final String TAG = MessageListViewStyle.class.getSimpleName();
 
     // Message text
-    private int messageTextSizeMine;
-    private int messageTextSizeTheirs;
-    private int messageTextColorMine;
-    private int messageTextColorTheirs;
-    private int messageTextStyleMine;
-    private int messageTextStyleTheirs;
+    public TextStyle messageTextMine;
+    public TextStyle messageTextTheirs;
+
     // Message Bubble
     private int messageBubbleDrawableMine;
     private int messageBubbleDrawableTheirs;
@@ -37,32 +34,15 @@ public class MessageListViewStyle extends BaseStyle {
     private int messageBorderWidthMine;
     private int messageBorderWidthTheirs;
     // Attachment
+    public TextStyle attachmentTitleTextMine;
+    public TextStyle attachmentTitleTextTheirs;
+
+    public TextStyle attachmentDescriptionTextMine;
+    public TextStyle attachmentDescriptionTextTheirs;
+    public TextStyle attachmentFileSizeText;
     private int attachmentBackgroundColorMine;
     private int attachmentBackgroundColorTheirs;
 
-    private int attachmentTitleTextSizeMine;
-    private int attachmentTitleTextColorMine;
-    private int attachmentTitleTextStyleMine;
-
-    private int attachmentTitleTextSizeTheirs;
-    private int attachmentTitleTextColorTheirs;
-    private int attachmentTitleTextStyleTheirs;
-
-    private int attachmentDescriptionTextSizeMine;
-    private int attachmentDescriptionTextColorMine;
-    private int attachmentDescriptionTextStyleMine;
-
-    private int attachmentDescriptionTextSizeTheirs;
-    private int attachmentDescriptionTextColorTheirs;
-    private int attachmentDescriptionTextStyleTheirs;
-
-    private int attachmentFileSizeTextSizeMine;
-    private int attachmentFileSizeTextColorMine;
-    private int attachmentFileSizeTextStyleMine;
-
-    private int attachmentFileSizeTextSizeTheirs;
-    private int attachmentFileSizeTextColorTheirs;
-    private int attachmentFileSizeTextStyleTheirs;
     // Reaction
     private boolean reactionEnabled;
     // ReactionView
@@ -80,9 +60,8 @@ public class MessageListViewStyle extends BaseStyle {
     private boolean messageDateShow;
 
     // Date Separator
-    private int dateSeparatorDateTextSize;
-    private int dateSeparatorDateTextStyle;
-    private int dateSeparatorDateTextColor;
+    public TextStyle dateSeparatorDateText;
+
     private int dateSeparatorLineColor;
     private int dateSeparatorLineWidth;
     private int dateSeparatorLineDrawable;
@@ -94,13 +73,20 @@ public class MessageListViewStyle extends BaseStyle {
                 R.styleable.MessageListView, 0, 0);
 
         // Message Text
-        messageTextSizeMine = a.getDimensionPixelSize(R.styleable.MessageListView_streamMessageTextSizeMine, getDimension(R.dimen.stream_message_text_font_size));
-        messageTextColorMine = a.getColor(R.styleable.MessageListView_streamMessageTextColorMine, Color.BLACK);
-        messageTextStyleMine = a.getInt(R.styleable.MessageListView_streamMessageTextStyleMine, Typeface.NORMAL);
+        messageTextMine = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamMessageTextSizeMine, getDimension(R.dimen.stream_message_text_font_size))
+                .color(R.styleable.MessageListView_streamMessageTextColorMine, Color.BLACK)
+                .font(R.styleable.MessageListView_streamMessageTextFontMineAssets, R.styleable.MessageListView_streamMessageTextFontMine)
+                .style(R.styleable.MessageListView_streamMessageTextStyleMine, Typeface.NORMAL)
+                .build();
 
-        messageTextSizeTheirs = a.getDimensionPixelSize(R.styleable.MessageListView_streamMessageTextSizeTheirs, getDimension(R.dimen.stream_message_text_font_size));
-        messageTextColorTheirs = a.getColor(R.styleable.MessageListView_streamMessageTextColorTheirs, Color.BLACK);
-        messageTextStyleTheirs = a.getInt(R.styleable.MessageListView_streamMessageTextStyleTheirs, Typeface.NORMAL);
+        messageTextTheirs = new TextStyle.Builder(a)
+                .size( R.styleable.MessageListView_streamMessageTextSizeTheirs, getDimension(R.dimen.stream_message_text_font_size))
+                .color(R.styleable.MessageListView_streamMessageTextColorTheirs, Color.BLACK)
+                .font(R.styleable.MessageListView_streamMessageTextFontTheirsAssets, R.styleable.MessageListView_streamMessageTextFontTheirs)
+                .style(R.styleable.MessageListView_streamMessageTextStyleTheirs, Typeface.NORMAL)
+                .build();
+
         // Message Bubble
         messageBubbleDrawableMine = a.getResourceId(R.styleable.MessageListView_streamMessageBubbleDrawableMine, -1);
         messageBubbleDrawableTheirs = a.getResourceId(R.styleable.MessageListView_streamMessageBubbleDrawableTheirs, -1);
@@ -124,31 +110,44 @@ public class MessageListViewStyle extends BaseStyle {
         messageBorderWidthMine = a.getDimensionPixelSize(R.styleable.MessageListView_streamMessageBorderWidthMine, getDimension(R.dimen.stream_message_stroke));
         messageBorderWidthTheirs = a.getDimensionPixelSize(R.styleable.MessageListView_streamMessageBorderWidthTheirs, getDimension(R.dimen.stream_message_stroke));
         // Attachment
+        attachmentTitleTextMine = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAttachmentTitleTextSizeMine, getDimension(R.dimen.stream_attach_title_text))
+                .color(R.styleable.MessageListView_streamAttachmentTitleTextColorMine, getColor(R.color.stream_attach_title_text))
+                .font(R.styleable.MessageListView_streamAttachmentTitleTextFontAssetsMine, R.styleable.MessageListView_streamAttachmentTitleTextFontMine)
+                .style(R.styleable.MessageListView_streamAttachmentTitleTextStyleMine, Typeface.BOLD)
+                .build();
+
+        attachmentTitleTextTheirs = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAttachmentTitleTextSizeTheirs, getDimension(R.dimen.stream_attach_title_text))
+                .color(R.styleable.MessageListView_streamAttachmentTitleTextColorTheirs, getColor(R.color.stream_attach_title_text))
+                .font(R.styleable.MessageListView_streamAttachmentTitleTextFontAssetsTheirs, R.styleable.MessageListView_streamAttachmentTitleTextFontTheirs)
+                .style(R.styleable.MessageListView_streamAttachmentTitleTextStyleTheirs, Typeface.BOLD)
+                .build();
+
         attachmentBackgroundColorMine = a.getColor(R.styleable.MessageListView_streamAttachmentBackgroundColorMine, messageBackgroundColorMine);
         attachmentBackgroundColorTheirs = a.getColor(R.styleable.MessageListView_streamAttachmentBackgroundColorTheirs, messageBackgroundColorTheirs);
-        attachmentTitleTextSizeMine = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentTitleTextSizeMine, getDimension(R.dimen.stream_attach_title_text));
-        attachmentTitleTextColorMine = a.getColor(R.styleable.MessageListView_streamAttachmentTitleTextColorMine, getColor(R.color.stream_attach_title_text));
-        attachmentTitleTextStyleMine = a.getInt(R.styleable.MessageListView_streamAttachmentTitleTextStyleMine, Typeface.BOLD);
 
-        attachmentTitleTextSizeTheirs = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentTitleTextSizeTheirs, getDimension(R.dimen.stream_attach_title_text));
-        attachmentTitleTextColorTheirs = a.getColor(R.styleable.MessageListView_streamAttachmentTitleTextColorTheirs, getColor(R.color.stream_attach_title_text));
-        attachmentTitleTextStyleTheirs = a.getInt(R.styleable.MessageListView_streamAttachmentTitleTextStyleTheirs, Typeface.BOLD);
+        attachmentDescriptionTextMine = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAttachmentDescriptionTextSizeMine, getDimension(R.dimen.stream_attach_description_text))
+                .color(R.styleable.MessageListView_streamAttachmentDescriptionTextColorMine, getColor(R.color.stream_gray_dark))
+                .font(R.styleable.MessageListView_streamAttachmentDescriptionTextFontAssetsMine, R.styleable.MessageListView_streamAttachmentDescriptionTextFontMine)
+                .style(R.styleable.MessageListView_streamAttachmentDescriptionTextStyleMine, Typeface.NORMAL)
+                .build();
 
-        attachmentDescriptionTextSizeMine = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentDescriptionTextSizeMine, getDimension(R.dimen.stream_attach_description_text));
-        attachmentDescriptionTextColorMine = a.getColor(R.styleable.MessageListView_streamAttachmentDescriptionTextColorMine, getColor(R.color.stream_gray_dark));
-        attachmentDescriptionTextStyleMine = a.getInt(R.styleable.MessageListView_streamAttachmentDescriptionTextStyleMine, Typeface.NORMAL);
+        attachmentDescriptionTextTheirs = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAttachmentDescriptionTextSizeTheirs, getDimension(R.dimen.stream_attach_description_text))
+                .color(R.styleable.MessageListView_streamAttachmentDescriptionTextColorTheirs, getColor(R.color.stream_gray_dark))
+                .font(R.styleable.MessageListView_streamAttachmentDescriptionTextFontAssetsTheirs, R.styleable.MessageListView_streamAttachmentDescriptionTextFontTheirs)
+                .style(R.styleable.MessageListView_streamAttachmentDescriptionTextStyleTheirs, Typeface.NORMAL)
+                .build();
 
-        attachmentDescriptionTextSizeTheirs = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentDescriptionTextSizeTheirs, getDimension(R.dimen.stream_attach_description_text));
-        attachmentDescriptionTextColorTheirs = a.getColor(R.styleable.MessageListView_streamAttachmentDescriptionTextColorTheirs, getColor(R.color.stream_gray_dark));
-        attachmentDescriptionTextStyleTheirs = a.getInt(R.styleable.MessageListView_streamAttachmentDescriptionTextStyleTheirs, Typeface.NORMAL);
+        attachmentFileSizeText = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAttachmentFileSizeTextSize, getDimension(R.dimen.stream_attach_file_size_text))
+                .color(R.styleable.MessageListView_streamAttachmentFileSizeTextColor, getColor(R.color.stream_gray_dark))
+                .font(R.styleable.MessageListView_streamAttachmentFileSizeTextFontAssets, R.styleable.MessageListView_streamAttachmentFileSizeTextFont)
+                .style(R.styleable.MessageListView_streamAttachmentFileSizeTextStyle, Typeface.BOLD)
+                .build();
 
-        attachmentFileSizeTextSizeMine = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentFileSizeTextSizeMine, getDimension(R.dimen.stream_attach_file_size_text));
-        attachmentFileSizeTextColorMine = a.getColor(R.styleable.MessageListView_streamAttachmentFileSizeTextColorMine, getColor(R.color.stream_gray_dark));
-        attachmentFileSizeTextStyleMine = a.getInt(R.styleable.MessageListView_streamAttachmentFileSizeTextStyleMine, Typeface.BOLD);
-
-        attachmentFileSizeTextSizeTheirs = a.getDimensionPixelSize(R.styleable.MessageListView_streamAttachmentFileSizeTextSizeTheirs, getDimension(R.dimen.stream_attach_file_size_text));
-        attachmentFileSizeTextColorTheirs = a.getColor(R.styleable.MessageListView_streamAttachmentFileSizeTextColorTheirs, getColor(R.color.stream_gray_dark));
-        attachmentFileSizeTextStyleTheirs = a.getInt(R.styleable.MessageListView_streamAttachmentFileSizeTextStyleTheirs, Typeface.BOLD);
         // Reaction
         reactionEnabled = a.getBoolean(R.styleable.MessageListView_streamReactionEnabled, true);
 
@@ -169,23 +168,33 @@ public class MessageListViewStyle extends BaseStyle {
         avatarBorderColor = a.getColor(R.styleable.MessageListView_streamAvatarBorderColor, Color.WHITE);
         avatarBackGroundColor = a.getColor(R.styleable.MessageListView_streamAvatarBackGroundColor, getColor(R.color.stream_gray_dark));
 
-        avatarInitialTextSize = a.getDimensionPixelSize(R.styleable.MessageListView_streamAvatarTextSize, getDimension(R.dimen.stream_channel_initials));
-        avatarInitialTextColor = a.getColor(R.styleable.MessageListView_streamAvatarTextColor, Color.WHITE);
-        avatarInitialTextStyle = a.getInt(R.styleable.MessageListView_streamAvatarTextStyle, Typeface.BOLD);
+        avatarInitialText = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamAvatarTextSize, getDimension(R.dimen.stream_channel_initials))
+                .color(R.styleable.MessageListView_streamAvatarTextColor, Color.WHITE)
+                .font(R.styleable.MessageListView_streamAvatarTextFontAssets, R.styleable.MessageListView_streamAvatarTextFont)
+                .style(R.styleable.MessageListView_streamAvatarTextStyle, Typeface.BOLD)
+                .build();
 
         // Read State
         showReadState = a.getBoolean(R.styleable.MessageListView_streamShowReadState, true);
         readStateAvatarWidth = a.getDimensionPixelSize(R.styleable.MessageListView_streamReadStateAvatarWidth, getDimension(R.dimen.stream_read_state_avatar_width));
         readStateAvatarHeight = a.getDimensionPixelSize(R.styleable.MessageListView_streamReadStateAvatarHeight, getDimension(R.dimen.stream_read_state_avatar_height));
-        readStateTextSize = a.getDimensionPixelSize(R.styleable.MessageListView_streamReadStateTextSize, getDimension(R.dimen.stream_read_state_text_size));
-        readStateTextColor = a.getColor(R.styleable.MessageListView_streamReadStateTextColor, Color.BLACK);
-        readStateTextStyle = a.getColor(R.styleable.MessageListView_streamReadStateTextStyle, Typeface.BOLD);
+
+        readStateText = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamReadStateTextSize, getDimension(R.dimen.stream_read_state_text_size))
+                .color(R.styleable.MessageListView_streamReadStateTextColor, Color.BLACK)
+                .font(R.styleable.MessageListView_streamReadStateTextFontAssets, R.styleable.MessageListView_streamReadStateTextFont)
+                .style(R.styleable.MessageListView_streamReadStateTextStyle, Typeface.BOLD)
+                .build();
 
         threadEnabled = a.getBoolean(R.styleable.MessageListView_streamThreadEnabled, true);
 
-        dateSeparatorDateTextSize = a.getDimensionPixelSize(R.styleable.MessageListView_streamDateSeparatorDateTextSize, getDimension(R.dimen.stream_date_separator_text));
-        dateSeparatorDateTextColor = a.getColor(R.styleable.MessageListView_streamDateSeparatorDateTextColor, getColor(R.color.stream_gray_dark));
-        dateSeparatorDateTextStyle = a.getInt(R.styleable.MessageListView_streamDateSeparatorDateTextStyle, Typeface.BOLD);
+        dateSeparatorDateText = new TextStyle.Builder(a)
+                .size(R.styleable.MessageListView_streamDateSeparatorDateTextSize, getDimension(R.dimen.stream_date_separator_text))
+                .color(R.styleable.MessageListView_streamDateSeparatorDateTextColor, getColor(R.color.stream_gray_dark))
+                .font(R.styleable.MessageListView_streamDateSeparatorDateTextFontAssets, R.styleable.MessageListView_streamDateSeparatorDateTextFont)
+                .style(R.styleable.MessageListView_streamDateSeparatorDateTextStyle, Typeface.BOLD)
+                .build();
 
         dateSeparatorLineWidth = a.getDimensionPixelSize(R.styleable.MessageListView_streamDateSeparatorLineWidth, getDimension(R.dimen.stream_date_separator_line_width));
         dateSeparatorLineColor = a.getColor(R.styleable.MessageListView_streamDateSeparatorLineColor, getColor(R.color.stream_gray_dark));
@@ -195,18 +204,6 @@ public class MessageListViewStyle extends BaseStyle {
         messageDateShow = a.getBoolean(R.styleable.MessageListView_streamMessageDateShow, true);
 
         a.recycle();
-    }
-
-    public int getMessageTextSize(boolean isMine) {
-        return isMine ? messageTextSizeMine : messageTextSizeTheirs;
-    }
-
-    public int getMessageTextColor(boolean isMine) {
-        return isMine ? messageTextColorMine : messageTextColorTheirs;
-    }
-
-    public int getMessageTextStyle(boolean isMine) {
-        return isMine ? messageTextStyleMine : messageTextStyleTheirs;
     }
 
     public int getMessageBubbleDrawable(boolean isMine) {
@@ -224,6 +221,11 @@ public class MessageListViewStyle extends BaseStyle {
     public int getMessageBottomRightCornerRadius(boolean isMine) {
         return isMine ? messageBottomRightCornerRadiusMine : messageBottomRightCornerRadiusTheirs;
     }
+    // Attachment
+
+    public int getAttachmentBackgroundColor(boolean isMine) {
+        return isMine ? attachmentBackgroundColorMine : attachmentBackgroundColorTheirs;
+    }
 
     public int getMessageBottomLeftCornerRadius(boolean isMine) {
         return isMine ? messageBottomLeftCornerRadiusMine : messageBottomLeftCornerRadiusTheirs;
@@ -233,56 +235,12 @@ public class MessageListViewStyle extends BaseStyle {
         return isMine ? messageBackgroundColorMine : messageBackgroundColorTheirs;
     }
 
-
     public int getMessageBorderColor(boolean isMine) {
         return isMine ? messageBorderColorMine : messageBorderColorTheirs;
     }
 
     public int getMessageBorderWidth(boolean isMine) {
         return isMine ? messageBorderWidthMine : messageBorderWidthTheirs;
-    }
-
-    // Attachment
-
-
-    public int getAttachmentBackgroundColor(boolean isMine) {
-        return isMine ? attachmentBackgroundColorMine : attachmentBackgroundColorTheirs;
-    }
-
-    public int getAttachmentTitleTextSize(boolean isMine) {
-        return isMine ? attachmentTitleTextSizeMine : attachmentTitleTextSizeTheirs;
-    }
-
-    public int getAttachmentTitleTextColor(boolean isMine) {
-        return isMine ? attachmentTitleTextColorMine : attachmentTitleTextColorTheirs;
-    }
-
-    public int getAttachmentTitleTextStyle(boolean isMine) {
-        return isMine ? attachmentTitleTextStyleMine : attachmentTitleTextStyleTheirs;
-    }
-
-    public int getAttachmentDescriptionTextSize(boolean isMine) {
-        return isMine ? attachmentDescriptionTextSizeMine : attachmentDescriptionTextSizeTheirs;
-    }
-
-    public int getAttachmentDescriptionTextColor(boolean isMine) {
-        return isMine ? attachmentDescriptionTextColorMine : attachmentDescriptionTextColorTheirs;
-    }
-
-    public int getAttachmentDescriptionTextStyle(boolean isMine) {
-        return isMine ? attachmentDescriptionTextStyleMine : attachmentDescriptionTextStyleTheirs;
-    }
-
-    public int getAttachmentFileSizeTextSize(boolean isMine) {
-        return isMine ? attachmentFileSizeTextSizeMine : attachmentFileSizeTextSizeTheirs;
-    }
-
-    public int getAttachmentFileSizeTextColor(boolean isMine) {
-        return isMine ? attachmentFileSizeTextColorMine : attachmentFileSizeTextColorTheirs;
-    }
-
-    public int getAttachmentFileSizeTextStyle(boolean isMine) {
-        return isMine ? attachmentFileSizeTextStyleMine : attachmentFileSizeTextStyleTheirs;
     }
 
     // Reaction Dialog
@@ -323,20 +281,7 @@ public class MessageListViewStyle extends BaseStyle {
         return threadEnabled;
     }
 
-
     // Date Separator
-
-    public int getDateSeparatorDateTextSize() {
-        return dateSeparatorDateTextSize;
-    }
-
-    public int getDateSeparatorDateTextStyle() {
-        return dateSeparatorDateTextStyle;
-    }
-
-    public int getDateSeparatorDateTextColor() {
-        return dateSeparatorDateTextColor;
-    }
 
     public int getDateSeparatorLineColor() {
         return dateSeparatorLineColor;